Create File_Deleter
This program only deleted files NOT directories
The start date and end date define the time period during which you want to search for files that have been modified. Any file that has been modified between these two dates will be considered. The start date marks the beginning of the period, and the end date marks the end of the period.